requirements.txt
alembic==1.14.0
asyncpg==0.30.0
bcrypt==4.2.1
fastapi==0.115.6
passlib==1.7.4
pydantic==2.10.4
pydantic-settings==2.7.0
PyJWT[crypto]==2.10.1
python-multipart==0.0.20
SQLAlchemy[asyncio]==2.0.36
uvicorn==0.34.0
